Add menu module, helper method DOM
Added menu module to encapsulate the content related to menu creation.
The module contains a function to create menu which returns all the
elements necessary to create a food item menu

Created a helper method for DOM name "setTextContents". The method
takens and array of elements and an array of strings. It applies
textContent content property all elements in elements array with
corresponding values in string array